Move procfs-specific code into optional module
Making procfs support optional lets us build a "pure" filtering
version of Laurel for debugging purposes or for running in
non-auditd-plugin setups.
