+++ src/lib.rs
@@ -5,6 +5,8 @@ pub mod label_matcher;
 pub mod logger;
 pub mod parser;
 pub mod proc;
+#[cfg(feature = "procfs")]
+pub mod procfs;
 pub mod quoted_string;
 pub mod rotate;
 pub mod sockaddr;
